Refactor App.js (add Home and Login pages)
